Initial inspect-datagent test
